Authorization route function and uuid usage

+++ api/src/controllers/NewsController.js
@@ -2,6 +2,9 @@ const NewsRepository = require("../repository/NewsRepository")
 const NewsConstants = require("../constants/NewsConstants")
 const ServerConstants = require("../constants/ServerConstants")
 const { validationResult } = require("express-validator")
+const decoded_user = require("../utils/verifyUserAuthenticated")
+const UserRepository = require("../repository/UserRepository")
+const ERole = require("../utils/ERole")
 
 class NewsController{
     async get_all(req, res){
@@ -61,13 +64,19 @@ class NewsController{
             if(!validation_errors.isEmpty()){
                 return res.status(400).json({msg: NewsConstants.INVALID_DATA})
             }
+            
+            const user_id = decoded_user(req.headers['authorization']).id;
+            const verify_user = await UserRepository.verify_id(user_id);
+            if(!verify_user){
+                return res.status(403).json({msg: ServerConstants.NOT_AUTHORIZED})
+            }
 
             let newsExists = await NewsRepository.verify_title(title)
             if(newsExists){
                 return res.status(406).json({msg: NewsConstants.ALREADY_EXISTS})
             }
 
-            await NewsRepository.create(title, category_id, body)
+            await NewsRepository.create(title, user_id, category_id, body)
             res.json({status: NewsConstants.CREATED_SUCCESS})
         }catch(e){
             res.status(500).json({msg: ServerConstants.INTERNAL_ERROR})
@@ -88,6 +97,11 @@ class NewsController{
                 return res.status(404).json({msg: NewsConstants.NOT_FOUND})
             }
 
+            const user = decoded_user(req.headers['authorization']);
+            if(user.role != ERole.ADMIN && newsExists.user_id !== user.id){
+                return res.status(403).json({msg: ServerConstants.NOT_AUTHORIZED})
+            }
+
             await NewsRepository.delete_by_id(id)
             res.status(200).json({status: NewsConstants.DELETED_SUCCESS})
         }catch(e){
@@ -98,25 +112,29 @@ class NewsController{
     async update(req, res){
         try{
             let {id} = req.params
-            let {title, category_id, body} = req.body
+            let {title, is_active, category_id, body} = req.body
 
             const validation_errors = validationResult(req)
             if(!validation_errors.isEmpty()){
                 return res.status(400).json({msg: NewsConstants.INVALID_DATA})
             }
-
+            
+            let verifyId = await NewsRepository.find_by_id(id)
+            if(!verifyId){
+                return res.status(404).json({msg: NewsConstants.NOT_FOUND})
+            }
+            
+            const user = decoded_user(req.headers['authorization']);
+            if(user.role != ERole.ADMIN && verifyId.user_id !== user.id){
+                return res.status(403).json({msg: ServerConstants.NOT_AUTHORIZED})
+            }
+            
             let newsExists = await NewsRepository.verify_title_by_id(id, title)
             if(newsExists){
                 return res.status(406).json({msg: NewsConstants.ALREADY_EXISTS})
             }
 
-            let verifyId = await NewsRepository.find_by_id(id)
-            
-            if(!verifyId){
-                return res.status(404).json({msg: NewsConstants.NOT_FOUND})
-            }
-
-            await NewsRepository.update(id, title, category_id, body)
+            await NewsRepository.update(id, title, is_active, category_id, body)
             res.json({status: NewsConstants.UPDATED_SUCCESS})
         }catch(e){
             res.status(500).json({msg: ServerConstants.INTERNAL_ERROR})
